make absen

+++ application/models/Guru_model.php
@@ -11,7 +11,18 @@ class Guru_model extends CI_Model
     public function hapus($nip)
     {
         $this->db->where('nip', $nip);
+        $data_kelas = $this->db->get('tbl_kelas')->row_array();
+
+        $this->db->where('nip', $nip);
+        $data_mapel = $this->db->get('tbl_mapel')->row_array();
+
+        if($data_kelas || $data_mapel){
+            return false;
+        }
+
+        $this->db->where('nip', $nip);
         $this->db->delete('tbl_guru');
+        return true;
     }
 
     public function tambah($data)
